Fixed a few warnings about size_t to int conversion.

+++ ElementXML/src/sml_ElementXMLImpl.cpp
@@ -60,28 +60,28 @@ static char const* kCommentEndString	= "-->" ;
 
 static char const* kEncodeHex = "bin_encoding=\"hex\"" ;
 
-static const int kLenLT   = strlen(kLT) ;
-static const int kLenGT   = strlen(kGT) ;
-static const int kLenAMP  = strlen(kAMP) ;
-static const int kLenQUOT = strlen(kQUOT) ;
-static const int kLenAPOS = strlen(kAPOS) ;
+static const int kLenLT   = (int)strlen(kLT) ;
+static const int kLenGT   = (int)strlen(kGT) ;
+static const int kLenAMP  = (int)strlen(kAMP) ;
+static const int kLenQUOT = (int)strlen(kQUOT) ;
+static const int kLenAPOS = (int)strlen(kAPOS) ;
 
-static const int kLenCDataStart = strlen(kCDataStart) ;
-static const int kLenCDataEnd   = strlen(kCDataEnd) ;
+static const int kLenCDataStart = (int)strlen(kCDataStart) ;
+static const int kLenCDataEnd   = (int)strlen(kCDataEnd) ;
 
-static const int kLenStartTagOpen  = strlen(kStartTagOpen) ;
-static const int kLenStartTagClose = strlen(kStartTagClose) ;
-static const int kLenEndTagOpen    = strlen(kEndTagOpen) ;
-static const int kLenEndTagClose   = strlen(kEndTagClose) ;
+static const int kLenStartTagOpen  = (int)strlen(kStartTagOpen) ;
+static const int kLenStartTagClose = (int)strlen(kStartTagClose) ;
+static const int kLenEndTagOpen    = (int)strlen(kEndTagOpen) ;
+static const int kLenEndTagClose   = (int)strlen(kEndTagClose) ;
 
-static const int kLenEquals		   = strlen(kEquals) ;
-static const int kLenSpace		   = strlen(kSpace) ;
-static const int kLenQuote		   = strlen(kQuote) ;
+static const int kLenEquals		   = (int)strlen(kEquals) ;
+static const int kLenSpace		   = (int)strlen(kSpace) ;
+static const int kLenQuote		   = (int)strlen(kQuote) ;
 
-static const int kLenEncodeHex	   = strlen(kEncodeHex) ;
+static const int kLenEncodeHex	   = (int)strlen(kEncodeHex) ;
 
-static const int kLenCommentStartString = strlen(kCommentStartString) ;
-static const int kLenCommentEndString = strlen(kCommentEndString) ;
+static const int kLenCommentStartString = (int)strlen(kCommentStartString) ;
+static const int kLenCommentEndString = (int)strlen(kCommentEndString) ;
 
 inline static char const* ConvertSpecial(char base)
 {
@@ -145,7 +145,7 @@ static char binaryVal(char c)
 static void HexCharsToBinary(char const* pHexString, char*& pBinaryBuffer, int& length)
 {
 	// This length should always be an exact multiple of 2
-	int stringLength = strlen(pHexString) ;
+	int stringLength = (int)strlen(pHexString) ;
 
 	// The output will be exactly half the length of the input
 	length = (stringLength+1)/2 ;
@@ -867,7 +867,7 @@ int	 ElementXMLImpl::GetCharacterDataLength() const
 	if (m_DataIsBinary)
 		return this->m_BinaryDataLength ;
 	else
-		return strlen(m_CharacterData)+1 ;
+		return (int)strlen(m_CharacterData)+1 ;
 }
 
 /*************************************************************
@@ -932,13 +932,13 @@ int ElementXMLImpl::DetermineXMLStringLength(bool includeChildren) const
 	// The comment
 	if (m_Comment)
 	{
-		len += kLenCommentStartString + strlen(m_Comment) + kLenCommentEndString ;
+		len += kLenCommentStartString + (int)strlen(m_Comment) + kLenCommentEndString ;
 	}
 
 	// The start tag
 	if (m_TagName)
 	{
-		len += kLenStartTagOpen + strlen(this->m_TagName) + kLenStartTagClose ;
+		len += kLenStartTagOpen + (int)strlen(this->m_TagName) + kLenStartTagClose ;
 	}
 	
 	// The character data
@@ -955,7 +955,7 @@ int ElementXMLImpl::DetermineXMLStringLength(bool includeChildren) const
 		}
 		else if (m_UseCData)
 		{
-			len += kLenCDataStart + strlen(m_CharacterData) + kLenCDataEnd ;
+			len += kLenCDataStart + (int)strlen(m_CharacterData) + kLenCDataEnd ;
 		}
 		else
 		{
@@ -971,7 +971,7 @@ int ElementXMLImpl::DetermineXMLStringLength(bool includeChildren) const
 		xmlStringConst val = mapIter->second ;
 
 		// The attribute name is an identifier and can't contain special chars
-		len += kLenSpace + strlen(att) + kLenEquals ;
+		len += kLenSpace + (int)strlen(att) + kLenEquals ;
 
 		// The value can contain special chars
 		len += kLenQuote + CountXMLLength(val) + kLenQuote ;
@@ -995,7 +995,7 @@ int ElementXMLImpl::DetermineXMLStringLength(bool includeChildren) const
 	// The end tag
 	if (m_TagName)
 	{
-		len += kLenEndTagOpen + strlen(this->m_TagName) + kLenEndTagClose ;
+		len += kLenEndTagOpen + (int)strlen(this->m_TagName) + kLenEndTagClose ;
 	}
 
 	return len ;
